requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.1
pytz==2020.1
sqlparse==0.4.1
django-cors-headers==3.5.0
django-environ==0.4.5
django-rest-swagger==2.2.0 # https://github.com/marcgibbons/django-rest-swagger
psycopg2==2.8.3 --no-binary psycopg2  # https://github.com/psycopg/psycopg2